+++ src/app/core/data/meta/meta-data.service.ts
@@ -1,49 +1,19 @@
 import { Injectable } from '@angular/core';
-import { Observable, of } from 'rxjs';
-import { map } from 'rxjs/operators';
-
-import { ApiService } from 'src/app/core/http/api/api.service';
-import { FilterTagResponse, FilterItemResponse } from 'src/app/core/http/api/api.interfaces';
-
-import { FilterList } from 'src/app/core/models/offer-filter';
+import { Observable} from 'rxjs';
+import { OfferPropertyList } from 'src/app/core/models/offer-properties';
+import { OfferService } from 'src/app/core/http/offer/offer.service';
 
 @Injectable({
   providedIn: 'root',
 })
 export class MetaDataService {
-  filterList: FilterList[] = undefined;
-  constructor(private apiService: ApiService) {}
+  constructor(private offerService: OfferService) {}
 
-  // //////////////////////////////////
-  // get FilterTags
-  // //////////////////////////////////
-
-  public getFilterTags(): Observable<FilterList[]> {
-    if (this.filterList == undefined) {
-      return this.apiService.getFilterTags().pipe(
-        map((data: FilterTagResponse) => {
-          const filterArray: FilterItemResponse[] = data.filter;
-          this.filterList = this.mapMetaDataToFilterLists_de(filterArray);
-          //console.log('Filterlist: ', this.filterList);
-          return this.filterList;
-        })
-      );
-    } else {
-      //console.log('Filterlist already loaded.');
-      return of(this.filterList);
-    }
+  public getOfferProperties(): Observable<Map<string, OfferPropertyList>> {
+    return this.offerService.getOfferProperties();
   }
 
-  private mapMetaDataToFilterLists_de(dataArray): FilterList[] {
-    return dataArray.map((item: FilterItemResponse) => {
-      const tempList = item.list.map((listItem) => {
-        return {
-          id: listItem.id,
-          identifier: listItem.identifier,
-          description: listItem.description.de,
-        };
-      });
-      return new FilterList(item.tag, tempList);
-    });
+  public getFilterTags(): Observable<Map<string, OfferPropertyList>> {
+    return this.offerService.getOfferProperties();
   }
 }