+++ src/app/core/models/offer-properties.ts
@@ -1,0 +1,14 @@
+// Offer Properties
+export type PropertyItem = {
+  id: number;
+  identifier: string;
+  description: string;
+};
+
+// Liste mit Typ
+export class OfferPropertyList {
+  constructor(
+    public type: string,
+    public list: PropertyItem[]
+  ) {}
+}